fix(api): Keep passing lambda events to getCurrentUser
