+++ packages/api/src/__tests__/runtime.test.ts
@@ -93,7 +93,52 @@ describe('buildCedarContext', () => {
         token: 'auth-provider=test; session=token-123',
       },
       {
-        event: request,
+        event: {
+          body: null,
+          headers: { cookie: 'auth-provider=test; session=token-123' },
+          httpMethod: 'GET',
+          isBase64Encoded: false,
+          multiValueHeaders: {
+            cookie: ['auth-provider=test; session=token-123'],
+          },
+          multiValueQueryStringParameters: null,
+          path: '/graphql',
+          pathParameters: null,
+          queryStringParameters: null,
+          requestContext: {
+            accountId: 'cedar',
+            apiId: 'cedar',
+            authorizer: undefined,
+            httpMethod: 'GET',
+            identity: {
+              accessKey: null,
+              accountId: null,
+              apiKey: null,
+              apiKeyId: null,
+              caller: null,
+              clientCert: null,
+              cognitoAuthenticationProvider: null,
+              cognitoAuthenticationType: null,
+              cognitoIdentityId: null,
+              cognitoIdentityPoolId: null,
+              principalOrgId: null,
+              sourceIp: '',
+              user: null,
+              userAgent: null,
+              userArn: null,
+            },
+            path: '/graphql',
+            protocol: 'HTTP/1.1',
+            requestId: 'cedar-request',
+            requestTimeEpoch: expect.any(Number),
+            resourceId: 'cedar',
+            resourcePath: '/graphql',
+            stage: '',
+          },
+          resource: '/graphql',
+          stageVariables: null,
+        },
+        request,
         context: undefined,
       },
     ])
